Moved implementation to Common
