+++ Source/Frontend/UI/UI_Extensions.cs
@@ -2320,54 +2320,6 @@ public static class JsonHelper
     }
 }
 
-// https://stackoverflow.com/a/11640700/10923568
-internal static class RandomExtensions
-{
-    //returns a uniformly random ulong between ulong.Min inclusive and ulong.Max inclusive
-    public static ulong NextULong(this Random rng)
-    {
-        byte[] buf = new byte[8];
-        rng.NextBytes(buf);
-        return BitConverter.ToUInt64(buf, 0);
-    }
-
-    //returns a uniformly random long between long.Min inclusive and long.Max inclusive
-    public static long NextLong(this Random rng)
-    {
-        byte[] buf = new byte[8];
-        rng.NextBytes(buf);
-        return BitConverter.ToInt64(buf, 0);
-    }
-
-    //returns a uniformly random long between Min and Max without modulo bias
-    public static long NextLong(this Random rng, long min, long max, bool inclusiveUpperBound = false)
-    {
-        ulong range = (ulong)(max - min);
-
-        if (inclusiveUpperBound)
-        {
-            if (range == ulong.MaxValue)
-            {
-                return rng.NextLong();
-            }
-            range++;
-        }
-
-        if (range <= 0)
-        {
-            throw new ArgumentOutOfRangeException(nameof(max), "Max must be greater than min when inclusiveUpperBound is false, and greater than or equal to when true");
-        }
-
-        ulong limit = ulong.MaxValue - ulong.MaxValue % range;
-        ulong r;
-        do
-        {
-            r = rng.NextULong();
-        } while (r > limit);
-        return (long)(r % range + (ulong)min);
-    }
-}
-
 /// <summary>
 /// Provides a generic collection that supports data binding and additionally supports sorting.
 /// See http://msdn.microsoft.com/en-us/library/ms993236.aspx
