+++ Source/Libraries/Common/Common.csproj
@@ -55,6 +55,7 @@
       <DependentUpon>LogConsoleForm.cs</DependentUpon>
     </Compile>
     <Compile Include="Logging.cs" />
+    <Compile Include="Extensions\RandomExtensions.cs" />
     <Compile Include="Objects\OperationResults.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="StaticTools.cs" />
