+++ Source/Libraries/CorruptCore/CorruptCore_Extensions.cs
@@ -1032,96 +1032,6 @@ namespace RTCV.CorruptCore
         }
     }
 
-    // https://stackoverflow.com/a/11640700/10923568
-    internal static class RandomExtensions
-    {
-        //returns a uniformly random ulong between ulong.Min inclusive and ulong.Max inclusive
-        public static ulong NextULong(this Random rng)
-        {
-            byte[] buf = new byte[8];
-            rng.NextBytes(buf);
-            return BitConverter.ToUInt64(buf, 0);
-        }
-
-        //returns a uniformly random ulong between ulong.Min and Max without modulo bias
-        public static ulong NextULong(this Random rng, ulong max, bool inclusiveUpperBound = false)
-        {
-            return rng.NextULong(ulong.MinValue, max, inclusiveUpperBound);
-        }
-
-        //returns a uniformly random ulong between Min and Max without modulo bias
-        public static ulong NextULong(this Random rng, ulong min, ulong max, bool inclusiveUpperBound = false)
-        {
-            ulong range = max - min;
-
-            if (inclusiveUpperBound)
-            {
-                if (range == ulong.MaxValue)
-                {
-                    return rng.NextULong();
-                }
-
-                range++;
-            }
-
-            if (range <= 0)
-            {
-                throw new ArgumentOutOfRangeException(nameof(max), "Max must be greater than min when inclusiveUpperBound is false, and greater than or equal to when true");
-            }
-
-            ulong limit = ulong.MaxValue - ulong.MaxValue % range;
-            ulong r;
-            do
-            {
-                r = rng.NextULong();
-            } while (r > limit);
-
-            return r % range + min;
-        }
-
-        //returns a uniformly random long between long.Min inclusive and long.Max inclusive
-        public static long NextLong(this Random rng)
-        {
-            byte[] buf = new byte[8];
-            rng.NextBytes(buf);
-            return BitConverter.ToInt64(buf, 0);
-        }
-
-        //returns a uniformly random long between long.Min and Max without modulo bias
-        public static long NextLong(this Random rng, long max, bool inclusiveUpperBound = false)
-        {
-            return rng.NextLong(long.MinValue, max, inclusiveUpperBound);
-        }
-
-        //returns a uniformly random long between Min and Max without modulo bias
-        public static long NextLong(this Random rng, long min, long max, bool inclusiveUpperBound = false)
-        {
-            ulong range = (ulong)(max - min);
-
-            if (inclusiveUpperBound)
-            {
-                if (range == ulong.MaxValue)
-                {
-                    return rng.NextLong();
-                }
-                range++;
-            }
-
-            if (range <= 0)
-            {
-                return 0;
-            }
-
-            ulong limit = ulong.MaxValue - ulong.MaxValue % range;
-            ulong r;
-            do
-            {
-                r = rng.NextULong();
-            } while (r > limit);
-            return (long)(r % range + (ulong)min);
-        }
-    }
-
     public static class ObjectCopier
     {
         public static T Clone<T>(T source)
