+++ Source/Libraries/CorruptCore/Corruption Engines/RTC_CustomEngine.cs
@@ -7,6 +7,7 @@
     using System.Linq;
     using System.Numerics;
     using System.Windows.Forms;
+    using RTCV.Common.CustomExtensions;
     using RTCV.NetCore;
 
     public static class RTC_CustomEngine
